Moving ContextFactory to org.enso.common
