+++ engine/runtime/src/main/java/org/enso/interpreter/EnsoLanguage.java
@@ -16,6 +16,7 @@ import java.io.PrintStream;
 import java.util.List;
 import java.util.Objects;
 import org.enso.common.LanguageInfo;
+import org.enso.common.RuntimeOptions;
 import org.enso.compiler.Compiler;
 import org.enso.compiler.context.InlineContext;
 import org.enso.compiler.context.LocalScope;
@@ -43,7 +44,6 @@ import org.enso.interpreter.runtime.tag.Patchable;
 import org.enso.interpreter.util.FileDetector;
 import org.enso.lockmanager.client.ConnectedLockManager;
 import org.enso.logger.masking.MaskingFactory;
-import org.enso.polyglot.RuntimeOptions;
 import org.enso.syntax2.Line;
 import org.enso.syntax2.Tree;
 import org.graalvm.options.OptionCategory;
